+++ core-api/src/main/java/com/imlinker/coreapi/core/auth/security/jwt/JwtTokenProvider.java
@@ -1,4 +1,4 @@
-package com.imlinker.coreapi.core.auth.jwt;
+package com.imlinker.coreapi.core.auth.security.jwt;
 
 import com.imlinker.coreapi.support.exception.FilterExceptionHandler;
 import com.imlinker.domain.common.Email;
