+++ core-api/src/main/java/com/imlinker/coreapi/core/auth/security/oauth2/CustomAccessDeniedHandler.java
@@ -1,4 +1,4 @@
-package com.imlinker.coreapi.core.auth.oauth2;
+package com.imlinker.coreapi.core.auth.security.oauth2;
 
 import com.imlinker.coreapi.support.exception.FilterExceptionHandler;
 import com.imlinker.error.ErrorType;
@@ -6,22 +6,23 @@ import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import java.io.IOException;
 import lombok.RequiredArgsConstructor;
-import org.springframework.security.core.AuthenticationException;
-import org.springframework.security.web.AuthenticationEntryPoint;
+import org.springframework.security.access.AccessDeniedException;
+import org.springframework.security.web.access.AccessDeniedHandler;
 import org.springframework.stereotype.Component;
 
 @Component
 @RequiredArgsConstructor
-public class CustomAuthenticationEntryPoint implements AuthenticationEntryPoint {
+public class CustomAccessDeniedHandler implements AccessDeniedHandler {
 
     private final FilterExceptionHandler filterExceptionHandler;
 
     @Override
-    public void commence(
+    public void handle(
             HttpServletRequest request,
             HttpServletResponse response,
-            AuthenticationException authException)
+            AccessDeniedException accessDeniedException)
             throws IOException {
-        filterExceptionHandler.sendErrorMessage(response, ErrorType.UNAUTHENTICATED, authException);
+        filterExceptionHandler.sendErrorMessage(
+                response, ErrorType.UNAUTHORIZED, accessDeniedException);
     }
 }
