+++ core-api/src/main/java/com/imlinker/coreapi/core/auth/security/oauth2/CustomOAuth2User.java
@@ -1,6 +1,6 @@
-package com.imlinker.coreapi.core.auth.oauth2;
+package com.imlinker.coreapi.core.auth.security.oauth2;
 
-import com.imlinker.coreapi.core.auth.oauth2.vendor.OAuthVendor;
+import com.imlinker.coreapi.core.auth.security.oauth2.vendor.OAuthVendor;
 import java.util.Collection;
 import java.util.Collections;
 import java.util.Map;
