+++ core-api/src/main/resources/oauth/application-oauth-dev.yml
@@ -12,6 +12,8 @@ spring:
             client-name: kakao
             scope:
               - account_email
+              - profile_nickname
+              - profile_image
         provider:
           kakao:
             authorization-uri: https://kauth.kakao.com/oauth/authorize
